Fix compilation warnings on x64.
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@59141 91177308-0d34-0410-b5e6-96231b3b80d8

+++ include/mvm/PrintBuffer.h
@@ -138,7 +138,7 @@ public:
   /// writes8 - Writes a int64.
   inline PrintBuffer *writeS8(sint64 v) {
     char buf[32];
-    sprintf(buf, "%lld", v);
+    sprintf(buf, "%lld", (long long int)v);
     return write(buf);
   }
